Add a --simulate option
And a --simulate command line option that works together with --install
and doesn't actually install anything, it only produces a list of
packages and SlackBuilds to be installed.

The -s option is now short for --simulate instead of --search. It
makes sense since search is the default action and you never really have
to specify it in the command line.

Also, the check_for_root has been moved from the command line parsing
section to inside the respective functions. Makes the command line
parsing section cleaner.